backend/requirements.txt
firebase-admin
google-genai
requests
